+++ Runtime/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("RealPlayTester.Tests.PlayMode")]
+[assembly: InternalsVisibleTo("Project.VerificationTests")]